+++ tests/cases.js
@@ -12,8 +12,8 @@ module.exports = [
 	'--test-entry base/b-router/test',
 	'--test-entry base/b-virtual-scroll/test',
 	'--test-entry base/b-sidebar/test',
-	'--test-entry form/b-button/test',
 	'--test-entry base/b-list/test',
+	'--test-entry form/b-button/test',
 	'--test-entry super/i-block/test',
 	'--test-entry core/dom/image/test',
 	'--test-entry core/dom/in-view/test'