+++ CHANGELOG.md
@@ -42,6 +42,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Added support of method invoking `super/i-block/modules/state/set`
 * Added a new associated type: `Items` `bList`
+* Added support of the default slot `bDummy`
 * Added support of interpolation of a data provider response:
   * `iInput`
   * `bList`